requirements.txt
flask==2.3.3
flask-cors==4.0.0
pydantic==2.4.2
langchain-core>=0.2.0,<0.3.0
langgraph==0.0.55
langchain-openai>=0.1.8
langchain-mcp-adapters>=0.1.0
python-dotenv==1.0.0